Added permission, corrected pylint-mistakes

+++ app/resources/director_resources.py
@@ -1,4 +1,8 @@
+"""
+Module for director resources
+"""
 from flask import request, jsonify
+from flask_login import current_user
 from flask_restful import Resource
 
 from app import db
@@ -9,7 +13,15 @@ director_schema = DirectorSchema()
 
 
 class DirectorListResource(Resource):
-    def get(self):
+    """
+    Resources for directors
+    """
+    @staticmethod
+    def get():
+        """
+        Get directors
+        :return: Response
+        """
         return jsonify([{
             'director_id': director.director_id,
             'first_name': director.first_name,
@@ -17,7 +29,12 @@ class DirectorListResource(Resource):
 
         } for director in Director.query.all()])
 
-    def post(self):
+    @staticmethod
+    def post():
+        """
+        Adds a director
+        :return:
+        """
         new_director = Director(
             first_name=request.json["first_name"],
             last_name=request.json["last_name"]
@@ -28,11 +45,26 @@ class DirectorListResource(Resource):
 
 
 class DirectorResource(Resource):
-    def get(self, director_id):
+    """
+    Resource for one director
+    """
+    @staticmethod
+    def get(director_id):
+        """
+        Get onr director
+        :param director_id: id of director
+        :return: JSON
+        """
         director = Director.query.get_or_404(director_id)
         return director_schema.dump(director)
 
-    def patch(self, director_id):
+    @staticmethod
+    def patch(director_id):
+        """
+        Updates a director
+        :param director_id: id of director
+        :return: JSON
+        """
         director = Director.query.get_or_404(director_id)
 
         if 'first_name' in request.json:
@@ -42,33 +74,23 @@ class DirectorResource(Resource):
         db.session.commit()
         return director_schema.dump(director)
 
-    def delete(self, director_id):
-        director = Director.query.get_or_404(director_id)
-        db.session.delete(director)
-        db.session.commit()
-        return '', 204
+    @staticmethod
+    def delete(director_id):
+        """
+        Delete director
+        :param director_id: id of director
+        :return: Response
+        """
+        if current_user.is_authenticated and current_user.is_superuser:
+            director = Director.query.get_or_404(director_id)
+            db.session.delete(director)
+            db.session.commit()
+            return jsonify({
+                "status": 204,
+                "reason": "Director was deleted"
+            })
 
-# from flask import jsonify, request
-#
-# from app import app, db
-# from app.models import Genre, Director
-# from app.schemas import genre_schema, director_schema
-#
-#
-# @app.route('/director/', methods=['GET'])
-# def get_directors():
-#     all_directors = Director.query.all()
-#     result = director_schema.dump(all_directors)
-#     return jsonify(result)
-#
-#
-# # adding a genre
-# @app.route('/director/', methods=['POST'])
-# def add_director():
-#     first_name = request.json['first_name']
-#     last_name = request.json['last_name']
-#     my_director = Director(first_name, last_name)
-#     db.session.add(my_director)
-#     db.session.commit()
-#
-#     return director_schema.jsonify(my_director)
+        return jsonify({
+                "status": 401,
+                "reason": "User is not admin"
+            })
