+++ app/resources/genre_resources.py
@@ -1,6 +1,8 @@
-import json
-
+"""
+Module for genre resources
+"""
 from flask import request, jsonify
+from flask_login import current_user
 from flask_restful import Resource
 
 from app import db
@@ -10,28 +12,27 @@ from app.schemas import GenreSchema
 genre_schema = GenreSchema()
 
 
-class GenreEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, Genre):
-            return obj.__dict__
-        return json.JSONEncoder.default(self, obj)
-
-
 class GenreListResource(Resource):
+    """
+    Resource for genres
+    """
     @staticmethod
     def get():
-        # genres = Genre.query.all()
-        # return json.dumps(genres, cls=GenreEncoder)
+        """
+        Get genres
+        :return: list of JSONs
+        """
         return [{
             'genre_id': genre.genre_id,
             'genre_title': genre.genre_title
         } for genre in Genre.query.all()]
-        # genres_list = []
-        # for genre in genres:
-        #     genres_list.append(genre.to_dict())
-        # return json.dumps(genres)
 
-    def post(self):
+    @staticmethod
+    def post():
+        """
+        Adds a genre
+        :return: JSON
+        """
         new_genre = Genre(
             genre_title=request.json['genre_title']
         )
@@ -41,11 +42,26 @@ class GenreListResource(Resource):
 
 
 class GenreResource(Resource):
-    def get(self, genre_id):
+    """
+    Resource for one genre
+    """
+    @staticmethod
+    def get(genre_id):
+        """
+        Get genre by genre_id
+        :param genre_id: id of genre
+        :return: JSON
+        """
         genre = Genre.query.get_or_404(genre_id)
         return genre_schema.dump(genre)
 
-    def patch(self, genre_id):
+    @staticmethod
+    def patch(genre_id):
+        """
+        Update genre by id
+        :param genre_id: id of genre
+        :return: JSON
+        """
         genre = Genre.query.get_or_404(genre_id)
 
         if 'genre_title' in request.json:
@@ -54,38 +70,22 @@ class GenreResource(Resource):
         db.session.commit()
         return genre_schema.dump(genre)
 
-    def delete(self, genre_id):
-        genre = Genre.query.get_or_404(genre_id)
-        db.session.delete(genre)
-        db.session.commit()
-        return '', 204
-
-# from flask import jsonify, request
-#
-# from app import app, db
-# from app.models import Genre
-# from app.schemas import genre_schema
-#
-#
-# @app.route('/genres/', methods=['GET'])
-# def get_genres():
-#     # all_genres = Genre.query.all()
-#     # return 'lalala'
-#     # result = genre_schema.dump(all_genres)
-#     # return jsonify(result)
-#     return jsonify([{
-#         'genre_id': genre.genre_id,
-#         'genre_title': genre.genre_title
-#     } for genre in Genre.query.all()])
-
-
-# adding a genre
-# @app.route('/genre/', methods=['POST'])
-# def add_genre():
-#     genre_title = request.json['genre_title']
-#
-#     my_genre = Genre(genre_title)
-#     db.session.add(my_genre)
-#     db.session.commit()
-#
-#     return genre_schema.jsonify(my_genre)
+    @staticmethod
+    def delete(genre_id):
+        """
+        Delete genre by id
+        :param genre_id: id of genre
+        :return: Response
+        """
+        if current_user.is_authenticated and current_user.is_superuser:
+            genre = Genre.query.get_or_404(genre_id)
+            db.session.delete(genre)
+            db.session.commit()
+            return jsonify({
+                "status": 204,
+                "reason": "Genre was deleted"
+            })
+        return jsonify({
+                "status": 401,
+                "reason": "User is not admin"
+            })
